Fixed minor compiler warnings.

+++ src/trfact.c
@@ -3,7 +3,7 @@
 #include "loops.h"
 #define MAX_FACT 20     // 20! is the largest factorial that fits in an llu
 
-unsigned long fact(unsigned long long n, unsigned long long tail)
+unsigned long long fact(unsigned long long n, unsigned long long tail)
 {
    if (n == 0)
    {
@@ -19,7 +19,7 @@ int main(void)
 {
     for (int l = 0; l < LOOPS; l++)
     {
-        unsigned long f = fact(MAX_FACT, 1);
+        unsigned long long f = fact(MAX_FACT, 1);
         printf("%llu\n", f);
     }
 }
